// model/games/enums/Games.ts
export enum GAMES {
  RED = 'red',
  BLUE = 'blue',
  YELLOW = 'yellow',
  GOLD = 'gold',
  SILVER = 'silver',
  CRYSTAL = 'crystal',
  RUBY = 'ruby',
  SAPPHIRE = 'sapphire',
  EMERALD = 'emerald',
  FIRE_RED = 'fire-red',
  LEAF_GREEN = 'leaf-green',
  DIAMOND = 'diamond',
  PEARL = 'pearl',
  PLATINUM = 'platinum',
  HEART_GOLD = 'heart-gold',
  SOUL_SILVER = 'soul-silver',
  BLACK = 'black',
  WHITE = 'white',
  BLACK_2 = 'black-2',
  WHITE_2 = 'white-2',
  X = 'x',
  Y = 'y',
  OMEGA_RUBY = 'omega-ruby',
  ALPHA_SAPPHIRE = 'alpha-sapphire',
  SUN = 'sun',
  MOON = 'moon',
  ULTRA_SUN = 'ultra-sun',
  ULTRA_MOON = 'ultra-moon',
  LETS_GO_PIKACHU = 'lets-go-pikachu',
  LETS_GO_EEVEE = 'lets-go-eevee',
  SWORD = 'sword',
  SHIELD = 'shield',
  BRILLIANT_DIAMOND = 'brilliant-diamond',
  SHINING_PEARL = 'shining-pearl',
  LEGENDS_ARCEUS = 'legends-arceus',
}